+++ lib_drawable_view/src/main/java/com/android/base/ui/drawable/DButton.kt
@@ -4,13 +4,13 @@ import android.content.Context
 import android.util.AttributeSet
 import androidx.annotation.AttrRes
 import androidx.appcompat.widget.AppCompatButton
-import com.android.base.ui.common.TextColorView
+import com.android.base.ui.common.TextColor
 
 class DButton @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     @AttrRes defStyleAttr: Int = android.R.attr.buttonStyle
-) : AppCompatButton(context, attrs, defStyleAttr), DrawableView, TextColorView {
+) : AppCompatButton(context, attrs, defStyleAttr), DrawableView, TextColor {
 
     private val codeDrawableHelper = CodeDrawableHelper(context, attrs, defStyleAttr)